+++ tests/Service/YamlFilesPathServiceTest.php
@@ -9,15 +9,26 @@ class YamlFilesPathServiceTest extends TestCase
     public function testFindAllYamlFilesInDir()
     {
         $testsDir = ['./tests/yamlFiles/unSorted/'];
-        $yamlFiles = YamlFilesPathService::getPathToYamlFiles($testsDir);
+        $yamlFiles = YamlFilesPathService::getPathToYamlFiles($testsDir, []);
 
         $this->assertCount(6, $yamlFiles);
+    }
+
+    public function testFindAllYamlFilesInDirExceptOfExcluded()
+    {
+        $testsDir = ['./tests/yamlFiles/'];
+        $yamlFiles = YamlFilesPathService::getPathToYamlFiles($testsDir, [
+            './tests/yamlFiles/sorted/config',
+            './tests/yamlFiles/unSorted/route',
+        ]);
+
+        $this->assertCount(9, $yamlFiles);
     }
 
     public function testFindFile()
     {
         $testsFile = ['./tests/yamlFiles/sorted/yaml-getting-started.yml'];
-        $yamlFiles = YamlFilesPathService::getPathToYamlFiles($testsFile);
+        $yamlFiles = YamlFilesPathService::getPathToYamlFiles($testsFile, []);
 
         $this->assertCount(1, $yamlFiles);
     }
@@ -26,7 +37,7 @@ class YamlFilesPathServiceTest extends TestCase
     {
         $pathToFile = './tests/yamlFiles/unSorted/yaml-getting-started.yml';
         $testsFile = [$pathToFile];
-        $yamlFiles = YamlFilesPathService::getPathToYamlFiles($testsFile);
+        $yamlFiles = YamlFilesPathService::getPathToYamlFiles($testsFile, []);
         $foundFile = reset($yamlFiles);
 
         $this->assertEquals($pathToFile, $foundFile);